Split theme parsing helpers
